Fix: MemberInfo -> MemberDto Inner class 로 변경
